refactor Stripe endpoint-related types to go in one lib file
