+++ jsapp/js/account/stripe.utils.ts
@@ -2,16 +2,16 @@ import {when} from 'mobx';
 
 import {ACTIVE_STRIPE_STATUSES} from 'js/constants';
 import envStore from 'js/envStore';
-import type {SubscriptionInfo} from 'js/account/subscriptionStore';
-import subscriptionStore from 'js/account/subscriptionStore';
 import type {
   BasePrice,
   ChangePlan,
-  Product,
   Checkout,
-} from 'js/account/stripe.api';
+  Product,
+  SubscriptionInfo,
+} from 'js/account/stripe';
+import subscriptionStore from 'js/account/subscriptionStore';
 import {notify} from 'js/utils';
-import {ChangePlanStatus} from 'js/account/stripe.api';
+import {ChangePlanStatus} from 'js/account/stripe';
 
 // check if the currently logged-in user has a paid subscription in an active status
 // promise returns a boolean, or `null` if Stripe is not active - we check for the existence of `stripe_public_key`
