+++ jsapp/js/components/usageContainer.tsx
@@ -1,7 +1,7 @@
 import classnames from 'classnames';
 import prettyBytes from 'pretty-bytes';
 import React from 'react';
-import {RecurringInterval} from 'js/account/stripe.api';
+import {RecurringInterval} from 'js/account/stripe';
 import Icon from 'js/components/common/icon';
 import styles from 'js/components/usageContainer.module.scss';
 import {USAGE_WARNING_RATIO} from 'js/constants';
